Add js files to module
